feat: Add ThemeBtn component for toggling theme mode
This commit adds a new component called ThemeBtn, which is responsible for toggling the theme mode of the application. It uses the useTheme hook to access the current theme mode and toggleTheme function. The component renders a checkbox input and a styled div that represents the toggle button. When the checkbox is checked, the theme mode is set to "dark", and when it is unchecked, the theme mode is set to "light". This addition improves the user experience by allowing users to easily switch between light and dark themes.