feat: Add Loader component for displaying a loading spinner
A new component called Loader has been added to display a loading spinner. This component is used to indicate that the page is loading and provide visual feedback to the user. The Loader component uses an SVG spinner animation and is positioned in the center of the screen using flexbox. This addition improves the user experience by providing a clear indication of ongoing loading processes.